+++ CHANGELOG.md
@@ -17,6 +17,11 @@ The minimum supported Rust version is now 1.34.0.
 
 ### Removed
 
+- The deprecated `singleton` constructors have been removed. Please use `unit`
+  instead.
+- The deprecated methods `Vector::chunks` and `Vector::chunks_mut` have been
+  removed in favour of `Vector::leaves` and `Vector::leaves_mut` respectively.
+  (#50)
 - `im::iter::unfold_mut` has been removed, as there's no meaningful difference
   between it and rust-std 1.34.0's `std::iter::from_fn` with a captured state
   varible.